One approach to address rf/renderizer#7
This generates a hash based on the strings in each output
config group, and uses that hash as a build directory. Changing an
output config will cause a new set of images to be rendered for only
that config in a new output directory -- unless you've already
built images for that config before, and the source images haven't
changed. So if you switch a config back, you don't have to re-render
all over again.